update to match working version
Due to a string of somewhat unconventional troubleshooting efforts,
several bugfixes have been tested in a (mostly) production environment
without ever getting checked into git. The result is that the version of
NEPI-Bot that is running on test hardware is working, but is ahead of
its corresponding version control. This commit fixes that by updating
all botfiles to match those running on the test hardware.